Fix some test Update dep
