+++ package.json
@@ -46,8 +46,8 @@
 	"license": "MIT",
 	"devDependencies": {
 		"@vue/test-utils": "1.0.0-beta.24",
-		"ajv": "6.5.2",
-		"autoprefixer": "9.1.1",
+		"ajv": "6.5.3",
+		"autoprefixer": "9.1.5",
 		"babel-core": "6.26.3",
 		"babel-eslint": "8.2.6",
 		"babel-loader": "7.1.5",
@@ -59,8 +59,8 @@
 		"bumped": "0.10.10",
 		"bumped-terminal": "0.7.5",
 		"chai": "4.1.2",
-		"cleave.js": "1.4.3",
-		"conventional-changelog-cli": "2.0.1",
+		"cleave.js": "1.4.4",
+		"conventional-changelog-cli": "2.0.5",
 		"coveralls": "3.0.2",
 		"cross-env": "5.2.0",
 		"css-loader": "0.28.11",
@@ -84,8 +84,8 @@
 		"jquery": "3.3.1",
 		"jsdom": "11.12.0",
 		"jsdom-global": "3.0.2",
-		"lodash": "4.17.10",
-		"lolex": "2.7.1",
+		"lodash": "4.17.11",
+		"lolex": "2.7.5",
 		"mocha": "5.2.0",
 		"mocha-generators": "2.0.0",
 		"mocha-loader": "1.1.3",
