+++ test/unit/specs/fields/fieldLabel.spec.js
@@ -1,5 +1,6 @@
 import { mount, createLocalVue } from "@vue/test-utils";
 
+import Vue from "vue";
 import FieldLabel from "src/fields/core/fieldLabel.vue";
 
 const localVue = createLocalVue();
@@ -14,7 +15,10 @@ function createField(data, methods) {
 				getValueFromOption: global.getValueFromOption
 			}
 		},
-		propsData: data
+		propsData: {
+			eventBus: new Vue(),
+			...data
+		}
 	});
 	if (methods) {
 		_wrapper.setMethods(methods);
