+++ test/unit/specs/fields/fieldSelectEx.spec.js
@@ -1,5 +1,6 @@
 import { mount, createLocalVue } from "@vue/test-utils";
 
+import Vue from "vue";
 import FieldSelectEx from "src/fields/optional/fieldSelectEx.vue";
 
 const localVue = createLocalVue();
@@ -15,7 +16,10 @@ function createField(data, methods) {
 				getValueFromOption: global.getValueFromOption
 			}
 		},
-		propsData: data
+		propsData: {
+			eventBus: new Vue(),
+			...data
+		}
 	});
 	if (methods) {
 		_wrapper.setMethods(methods);
