+++ test/unit/specs/fields/fieldVueMultiSelect.spec.js
@@ -17,7 +17,10 @@ function createField(data, methods) {
 				getValueFromOption: global.getValueFromOption
 			}
 		},
-		propsData: data,
+		propsData: {
+			eventBus: new Vue(),
+			...data
+		},
 		components: {
 			multiselect: VueMultiSelect
 		}
